restructure file hierarchy and rename previous findCoordinates files
