+++ findCoordsDemo.html
@@ -58,11 +58,15 @@
 
     function findCoordinates() {
       var startPoint = new google.maps.LatLng(51.5178767, -0.0762007);
-      var distance = 100;
-      var bearing = 0;
+      var distance = 0;
+      var heading = 0;
 
-      var endPoint = google.maps.geometry.spherical.computeOffset(startPoint, distance, bearing);
-      console.log(endPoint);
+      var endPoint = google.maps.geometry.spherical.computeOffset(startPoint, distance, heading);
+      console.log(endPoint.lat());
+      console.log(endPoint.lng());
+      console.log(endPoint.toJSON());
+      console.log(endPoint.toString());
+      console.log(endPoint.toUrlValue());
       var waypointB = new google.maps.Marker({
         position: endPoint,
         map: map
