+++ server/src/findCoordinatesWithMaths.js
